extract UpsellTeamPro
